requirements.txt
# Python 3.10+ recommended
# Install with: pip install -r requirements.txt

requests>=2.31.0
beautifulsoup4>=4.12.0
spotipy>=2.23.0
rapidfuzz>=3.0.0
python-dotenv>=1.0.0
